Extract validatePhotoUrls middleware, fix IPv6 URL validation, add auth factory for testability
